requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-telegram-bot==21.4

# OpenAI (Sora 2 + GPT-4 + AgentKit)
openai==1.54.0

# Database
supabase==2.7.4
psycopg2-binary>=2.9.11

# Background Tasks
apscheduler==3.10.4
celery[redis]==5.4.0

# HTTP & APIs
httpx==0.27.0
aiohttp==3.10.5
requests==2.32.4

# Social Media APIs & Scrapers
tweepy==4.14.0  # Twitter/X
instagrapi==2.2.1  # Instagram
TikTokApi==7.2.0  # TikTok
beautifulsoup4==4.14.2  # HTML parsing
lxml==6.0.2  # XML/HTML parser
playwright==1.55.0  # Browser automation for scraping

# Video Processing
ffmpeg-python==0.2.0
Pillow==10.4.0

# Data & Validation
pydantic>=2.7.1,<3.0.0
pydantic-settings==2.5.2

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.1

# Monitoring & Logging
sentry-sdk==2.15.0
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
